:sparkles: ext to format string into query param and some adjusts on BO file

+++ app/src/main/java/br/com/likeflix/domain/SearchedMoviesBO.kt
@@ -5,24 +5,24 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class SearchedMoviesBO(
-    val page: Double,
-    val total_results: Double,
-    val total_pages: Double,
+    val page: Long,
+    val total_results: Long,
+    val total_pages: Long,
     val results: List<MovieBO>
 ): Parcelable
 
 @Parcelize
 data class MovieBO(
     val popularity: Double? = null,
-    val vote_count: Double? = null,
+    val vote_count: Long? = null,
     val video: Boolean? = null,
     val poster_path: String,
-    val id: Double,
+    val id: Long,
     val adult: Boolean,
     val backdrop_path: String? = null,
     val original_language: String,
     val original_title: String,
-    val genre_ids: List<Double>,
+    val genre_ids: List<Long>,
     val title: String,
     val vote_average: Double? = null,
     val overview: String,
